Update SigninForm component and App component styles
