+++ client/src/App.css
@@ -1,0 +1,3 @@
+html {
+  background-color: rgb(15, 15, 17);
+}
