+++ client/src/components/SignInLanding/SignInLanding.jsx
@@ -1,0 +1,19 @@
+function SignInLanding() {
+  return (
+    <div className="max-w-md">
+      <span className="text-lg text-blue-400 font-bold">
+        Sign in to your Account
+      </span>
+      <h2 className="mt-8 mb-12 text-5xl font-bold font-heading text-white">
+        Sign in your account to keep track of your scores and receive updates on
+        more quizzes
+      </h2>
+      <p className="text-lg text-gray-400">
+        <span>repetition repetition repetition </span>
+        <span className="text-white"> is the key to success</span>
+      </p>
+    </div>
+  );
+}
+
+export default SignInLanding;
